Fixed compile errors.

+++ Server/src/petriwars/types/Point.java
@@ -1,8 +1,8 @@
 package petriwars.types;
 
 public class Point {
-	float x, y;
-	public Point(float x, float y){
-		this.x=x; this.y=y;
-	}
+ double x, y;
+ public Point(double x, double y){
+  this.x=x; this.y=y;
+ }
 }
